New Entity structure, builder, and refactors
