--- src/Engine/Components/ComponentTemplate.rs
@@ -1,43 +1,0 @@
-
-use crate::Engine::GameEntity::Entity;
-use crate::Engine::Components::Component;
-use crate::Engine::Frame::GameFrame;
-use crate::Engine::GameAPI::GameAPI;
-
-
-use std::sync::{RwLock, Mutex, Arc};
-use std::rc::Rc;
-
-/// General template for creating new components
-pub struct ComponentTemplate
-{
-}
-
-impl ComponentTemplate
-{
-    pub fn Create() -> Rc<RwLock<Self>>
-    {
-        Rc::new(
-            RwLock::new(
-                Self
-                {
-                }
-            )
-        )
-    }
-}
-
-impl Component for ComponentTemplate
-{
-    fn start(&mut self, 
-        entity: &mut Entity, 
-        api: Arc<Mutex<GameAPI>>)
-    {
-        // Add start life-time
-    }
-
-    fn update(&mut self, entity: &mut Entity, frame: &GameFrame, api: Arc<Mutex<GameAPI>>)
-    {
-        // Add update logic
-    }
-}
