+++ src/GrapplerSceneBuilder.rs
@@ -5,12 +5,14 @@ use glium::glutin::surface::WindowSurface;
 use crate::Engine::Audio::sample::{EAudioSpace, ETargetTrack};
 use crate::Engine::Collision::collider::{ECollisionTag, ECollisionType};
 use crate::Engine::Components::AudioSource::AudioPlayer;
-use crate::Engine::Components::Collider;
+use crate::Engine::Components::{AudioSource, Collider};
 use crate::Engine::Components::RenderComponents::{Renderer2D, Sprite};
-use crate::Engine::GameEntity::Entity;
+use crate::Engine::GameEntity::TEntity;
 use crate::Engine::Math::Float3;
 use crate::Engine::SceneBuilder::{Scene, TSceneBuilder};
-use crate::{LunaController, Grappler};
+use crate::{LunaController, Boxer};
+use crate::Engine::Tile::Tile;
+use crate::Entities::Entities;
 
 pub struct GCSBSceneBuilder
 {
@@ -19,17 +21,17 @@ pub struct GCSBSceneBuilder
 impl TSceneBuilder for GCSBSceneBuilder
 {
     fn LoadScene(name: String, rawScene: String, display: &Display<WindowSurface>)
-        -> Vec<Rc<RefCell<Entity>>>
+        -> Entities
     {
         println!("Loaded Scene: {}", name);
 
         let lines = rawScene.lines().filter( |x| !x.contains("*"));
 
-        lines
-            .map( |x| Self::ParseEntity(x, display))
-            .filter(|x| x.is_some())
-            .map(|x| x.unwrap())
-            .collect()
+        let mut entities = Entities::Create();
+
+        lines.for_each( |x| Self::ParseEntity(x, &mut entities, display));
+
+        entities
     }
 }
 
@@ -40,7 +42,7 @@ impl GCSBSceneBuilder
         Self{}
     }
 
-    fn ParseEntity(entry: &str, display: &Display<WindowSurface>) -> Option<Rc<RefCell<Entity>>>
+    fn ParseEntity(entry: &str, entities: &mut Entities, display: &Display<WindowSurface>)
     {
         let mut tokens : Vec<String> = Vec::new();
         entry
@@ -51,11 +53,11 @@ impl GCSBSceneBuilder
 
         match objectType.as_str()
         {
-            "Player" => Some(Self::BuildPlayer(tokens, display)),
-            "Tile" => Some(Self::BuildTile(tokens, display)),
-            "Audio" => Some(Self::BuildAudioSource(tokens, display)),
-            _ => None
-        }
+            "Player" => Self::BuildPlayer(tokens, entities, display),
+            "Tile" => Self::BuildTile(tokens, entities, display),
+            "Audio" => Self::BuildAudioSource(tokens, entities, display),
+            _ => {}
+        };
     }
 
     /// constructs a player object
@@ -63,7 +65,7 @@ impl GCSBSceneBuilder
     /// Entry Structure:
     ///     1 - name
     ///     2 - position
-    fn BuildPlayer(data: Vec<String>, display: &Display<WindowSurface>) -> Rc<RefCell<Entity>>
+    fn BuildPlayer(data: Vec<String>, entities: &mut Entities, display: &Display<WindowSurface>)
     {
         let name = data[1].as_str();
         let position = Float3::FromString(data[2].as_str());
@@ -75,39 +77,12 @@ impl GCSBSceneBuilder
             isLit = data[3].as_str().parse().unwrap();
         }
 
-        let player = Rc::new(RefCell::new(Entity::Create(name, position)));
+        let player = Boxer::Boxer::Create(
+            32.0,
+            display,
+            position);
 
-        let renderComponent =
-            Renderer2D::New(
-                &display,
-                Sprite::new(
-                    "Assets/run_down.png",
-                    &display,
-                    4,
-                    (2,2),
-                    0.001),
-                isLit
-                                
-            );
-
-        let movementComponent = Grappler::GrapplerController::new(16.0f32, &display);
-        //LunaController::LunaController::new(16.0f32, &display);
-
-        let mut playerMut = player.borrow_mut();
-        playerMut.add_component(movementComponent);
-        playerMut.add_component(renderComponent);
-        playerMut.add_component(
-            Collider::Collider::Create(
-                display.clone(),
-                position,
-                Float3::new(32.0, 32.0, 32.0),
-                ECollisionType::Solid,
-                ECollisionTag::None
-            ));
-
-        drop(playerMut);
-
-        player
+        entities.Boxer.push(player);
     }
 
     /// Builds a static tile object
@@ -118,7 +93,7 @@ impl GCSBSceneBuilder
     ///     3 - asset path
     ///     4 - is a collider
     ///     5 - tag
-    fn BuildTile(data: Vec<String>, display: &Display<WindowSurface>) -> Rc<RefCell<Entity>>
+    fn BuildTile(data: Vec<String>, entities: &mut Entities,  display: &Display<WindowSurface>)
     {
         // 1 - name
         let name = data[1].as_str();
@@ -130,7 +105,7 @@ impl GCSBSceneBuilder
         let assetPath = data[3].as_str();
 
         // 4 - is a collider - TODO
-        let mut collider = false;
+        let mut hasCollider = false;
 
         if data.len() >= 5
         {
@@ -138,51 +113,52 @@ impl GCSBSceneBuilder
 
             if(o.is_ok())
             {
-                collider = data[4].as_str().parse().unwrap();
+                hasCollider = data[4].as_str().parse().unwrap();
 
             }
         }
 
         // 5 - is lit
-
         let mut isLit = false;
         if data.len() >= 6
         {
             isLit = data[5].as_str().parse().unwrap();
         }
 
-        let tile = Rc::new(RefCell::new(Entity::Create(name, position)));
+        let mut collider = None;
+        if hasCollider
+        {
+            collider = Some(
+                Collider::Collider::Create(
+                    display.clone(),
+                    position,
+                    Float3::new(32.0, 32.0, 1.0),
+                    ECollisionType::Solid,
+                    ECollisionTag::None
+                )
+            );
+        }
+        else
+        {
+            collider = None;
+        }
 
-        let renderComponent =
-            Renderer2D::New(
-                &display,
+        let tile =
+            Tile::Create(
+                name,
+                position,
                 Sprite::new(
                     assetPath,
                     &display,
                     1,
                     (1,1),
                     0.001),
-                isLit
+                isLit,
+                display,
+                collider
             );
 
-        let mut tileMut = tile.borrow_mut();
-        tileMut.add_component(renderComponent);
-
-        if collider
-        {
-            tileMut.add_component(
-                Collider::Collider::Create(
-                    display.clone(),
-                    position,
-                    Float3::new(32.0, 32.0, 32.0),
-                    ECollisionType::Solid,
-                    ECollisionTag::None
-                ));
-        }
-
-        drop(tileMut);
-
-        tile
+        entities.Tiles.push(tile);
     }
 
     /// Constructs an audio source object.
@@ -195,8 +171,7 @@ impl GCSBSceneBuilder
     ///     5 - space(2D/ 3D)
     ///     6 - track(music/ sfx)
     ///
-    fn BuildAudioSource(data: Vec<String>, display: &Display<WindowSurface>)
-                        -> Rc<RefCell<Entity>>
+    fn BuildAudioSource(data: Vec<String>, entities: &mut Entities, display: &Display<WindowSurface>)
     {
         let name = data[1].as_str();
         let position = Float3::FromString(data[2].as_str());
@@ -206,12 +181,10 @@ impl GCSBSceneBuilder
         let volume : f32 = data[4].as_str().trim().parse().unwrap();
 
         // TODO: implement remaining properties
-        let audioSource = Rc::new(RefCell::new(Entity::Create(name, position)));
-
-        let mut mutEnt = audioSource.borrow_mut();
-
-        let audioSourceComp =
+        let audioSource =
             AudioPlayer::Create(
+                name,
+                position,
                 assetPath,
                 volume,
                 true,
@@ -219,9 +192,6 @@ impl GCSBSceneBuilder
                 ETargetTrack::Music
             );
 
-        mutEnt.add_component(audioSourceComp);
-        drop(mutEnt);
-
-        audioSource
+        entities.AudioPlayers.push(audioSource);
     }
 }
